Update transifex usage
This uses force mode to download po files, always. Normal operation is
to use the file modification date for comparison, which doesn't work
because of Git.

This also sets `onlyreviewed` mode, which will omit strings which are
not translated yet. This will result in the underlying source
translation being used in place, until the string is reviewed in
transifex.